refactor: 채팅방 프리뷰, 채팅방 열기 기능을 ChatRoom 관련 책임으로 분리
